+++ src/Article.js
@@ -38,24 +38,6 @@ const getLastEdit = (name, action) =>
 			}
 		})
 
-const thumbnailWidth = '320',
-	fullWidth = '800'
-
-export const imageResizer = (size) => (src) =>
-	src.includes('imgur.com')
-		? src.replace(/\.(png|jpg)$/, size + '.jpg')
-		: src.includes('unsplash.com')
-			? src.replace(
-				/w=[0-9]+\&/,
-				(_, p1) => `w=${size === 'm' ? thumbnailWidth : fullWidth}&`
-			)
-			: src.includes('medium.com')
-				? src.replace(
-					/max\/[0-9]+\//,
-					(_, p1) => `max/${size === 'm' ? thumbnailWidth : fullWidth}/`
-				)
-				: src
-
 export default ({}) => {
 	const {id} = useParams()
 	console.log(id, parsedArticles)
