+++ src/utils.js
@@ -11,6 +11,26 @@ function RouterLink(props) {
 		)
 }
 
+
+const thumbnailWidth = '320',
+	fullWidth = '800'
+
+export const imageResizer = (size) => (src) =>
+	src.includes('imgur.com')
+		? src.replace(/\.(png|jpg)$/, size + '.jpg')
+		: src.includes('unsplash.com')
+			? src.replace(
+				/w=[0-9]+\&/,
+				(_, p1) => `w=${size === 'm' ? thumbnailWidth : fullWidth}&`
+			)
+			: src.includes('medium.com')
+				? src.replace(
+					/max\/[0-9]+\//,
+					(_, p1) => `max/${size === 'm' ? thumbnailWidth : fullWidth}/`
+				)
+				: src
+
+
 export const Markdown = ({source}) => <ReactMarkdown
 	renderers={{image: ImageRenderer, link: RouterLink}}
 	source={source}
